src/graph-data-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-core==1.29.0
pydantic>=2.7.0
python-dotenv==1.0.0
neo4j==5.14.1
langgraph==0.0.32
langchain==0.1.0
langchain-core==0.1.10
